Summarize the result for powerful graph fused lasso inference

Summarize the result for powerful graph fused lasso inference

summary.fusedlasso_inf(object, ...)

Arguments

object
output from running spike_estimates

...
to be passed to methods

Value

A data frame with summarized results

Examples

lev1 <- 0 # mean for group 1
lev2 <- 3 # mean (absolute value) for group 2/3
sigma <- 1 # level of noise
nn <- 8 # grid size
Dmat <- graph2D_Dmat(nn^2) # generate D matrix for the 2D fused lasso
### Create the underlying signal
A <- matrix(lev1, ncol=nn, nrow = nn)
A[1:round(nn/3),1:round(nn/3)] <- 1*lev2
A[(nn-2):(nn),(nn-2):(nn)] <- -1*lev2
set.seed(2005)
A.noisy <- A + rnorm(nn^2,mean=0,sd=sigma)
y <- do.call(c,lapply(1:nrow(A.noisy),function(irow)A.noisy[irow,]))
### Run a test for a difference in means between estimated clusters 1 and 2
result_demo <- fusedlasso_inf(y=y, D=Dmat, c1=1, c2=2, method="K", sigma=sigma, K=13)
summary(result_demo)

#>   cluster_1 cluster_2 diff_in_means    pval_c1c2 pval_hyun LCB UCB
#> 1         1         2     -6.515453 6.854138e-30 0.1772895  NA  NA